Correcting unit test to conform with AddessedPhoneNumber translation

+++ notification-core/src/main/java/fr/sii/notification/sms/message/addressing/NumberingPlanIndicator.java
@@ -8,29 +8,29 @@ package fr.sii.notification.sms.message.addressing;
  */
 public enum NumberingPlanIndicator {
 	/** Unknown = 0. */
-	UNKNOWN(0),
+	UNKNOWN((byte) 0),
 	/** ISDN/telephone numbering plan (E163/E164) = 1. */
-	ISDN_TELEPHONE(1),
+	ISDN_TELEPHONE((byte) 1),
 	/** Data numbering plan (X.121) = 3. */	
-	DATA(3),
+	DATA((byte) 3),
 	/** Telex numbering plan (F.69) = 4. */	
-	TELEX(4),
+	TELEX((byte) 4),
 	/** Land Mobile (E.212) =6. */
-	LAND_MOBILE(6),
+	LAND_MOBILE((byte) 6),
 	/** National numbering plan = 8. */
-	NATIONAL(8),
+	NATIONAL((byte) 8),
 	/** Private numbering plan = 9. */
-	PRIVATE(9),
+	PRIVATE((byte) 9),
 	/** ERMES numbering plan (ETSI DE/PS 3 01-3) = 10. */
-	ERMES(10),
+	ERMES((byte) 10),
 	/** Internet (IP) = 13. */
-	IP(13),
+	IP((byte) 13),
 	/** WAP Client Id (to be defined by WAP Forum) = 18. */
-	WAP(13);
+	WAP((byte) 13);
 
-	private int value;
+	private byte value;
 
-	private NumberingPlanIndicator(int value) {
+	private NumberingPlanIndicator(byte value) {
 		this.value = value;
 	}
 
@@ -39,7 +39,7 @@ public enum NumberingPlanIndicator {
 	 * 
 	 * @return the byte value.
 	 */
-	public int value() {
+	public byte value() {
 		return value;
 	}
 
@@ -54,7 +54,7 @@ public enum NumberingPlanIndicator {
 	 *             if there is no enum const associated with specified int
 	 *             value.
 	 */
-	public static NumberingPlanIndicator valueOf(int value) {
+	public static NumberingPlanIndicator valueOf(byte value) {
 		for (NumberingPlanIndicator current : values()) {
 			if (current.value == value)
 				return current;
